Restructures scaling.py into a scaling module.
